Make x64 jmp target a u32

+++ src/x64_gen/lir_to_x64.h
@@ -140,7 +140,7 @@ typedef struct X64__Instr {
         } pop;
 
         struct {
-            size_t target;
+            u32 target;
         } jmp; // Also for JMP_TO_RET, jb, jl, je, ...
 
         struct {
